requirements.txt
playwright>=1.40.0
anthropic>=0.39.0
gspread>=6.0.0
google-auth>=2.20.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
